Move ConfigModule to a new `guice` subproject.
