+++ core/build.sbt
@@ -3,7 +3,6 @@ import Dependencies._
 name := "common-core"
 
 libraryDependencies ++= Seq(
-  scalaGuice,
   sprayJson,
   typesafeConfig,
   Logging.logbackClassic,
