+++ guice/build.sbt
@@ -1,0 +1,8 @@
+import Dependencies._
+
+name := "common-guice"
+
+libraryDependencies ++= Seq(
+  scalaGuice,
+  typesafeConfig
+)
