+++ guice/src/test/resources/org/allenai/common/guice/module.conf
