+++ guice/src/test/resources/org/allenai/common/guice/test_rename.conf
